测试

+++ src/main/java/com/xiliulou/electricity/entity/EleUserAuth.java
@@ -27,7 +27,7 @@ public class EleUserAuth {
     /**
     * 资料项id
     */
-    private Long entryId;
+    private Integer entryId;
     /**
     * 资料项内容
     */
